refactor(wallet): extract Client::Crypto concern module (#576)
